requirements.txt
blinker==1.4
Bootstrap-Flask==1.2.0
Click==7.0
Faker==3.0.0
Flask==1.1.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Sitemap==0.3.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mysql-connector==2.2.9
Pillow==6.2.1
python-dateutil==2.8.1
python-dotenv==0.10.3
six==1.13.0
SQLAlchemy==1.3.12
text-unidecode==1.3
Werkzeug==0.16.0
WTForms==2.2.1
